==> 06_GeoBoundaries.ipynb <==
pip install geopandas

from bs4 import BeautifulSoup
import pandas as pd
import re
import requests
import os
import geopandas as gpd

url_cnty = r'https://www2.census.gov/geo/tiger/TIGER2019/CBSA/tl_2019_us_cbsa.zip'
cnty = gpd.read_file(url_cnty)

def get_zip(url):
    front_page = requests.get(url,verify=False)
    soup = BeautifulSoup(front_page.content,'html.parser')
    zf = soup.find_all("a",href=re.compile(r"zip"))
    # Maybe should use href
    zl = [os.path.join(url,i['href']) for i in zf]
    return zl

base_url = r'https://www2.census.gov/geo/tiger/TIGER2020/PLACE/'
res = get_zip(base_url)

geo_place = []
for surl in res:
    geo_place.append(gpd.read_file(surl))

geo_full = pd.concat(geo_place)

geo_full.to_file("../tl_2021_us_place.geojson", driver="GeoJSON")

cnty = gpd.read_file("/content/drive/MyDrive/G00-place/geo_tiger_bdry/US/tl_2021_us_county/tl_2021_us_county.shp")
cnty

cities_cnty = sent_places.sjoin(cnty, how="right")
cities_cnty

### Reference

https://andrewpwheeler.com/2022/02/28/downloading-geo-files-from-census-ftp-using-python/
